added shared contents and fft runtime folders, started dealing with fft inplementation. Ommited data filtering for now, moving from floating point operations to non floating point operations

+++ InteractiveRaveLights/adc_runtime/adc_runtime_main.h
@@ -4,6 +4,6 @@
 #include "stdio.h"
 
 void adc_runtime_thread(void *ptr);
-void adc_copy_filtered_data(float *input_buffer, size_t input_buffer_size);
+void adc_copy_filtered_data(int16_t *input_buffer, size_t input_buffer_size);
 size_t adc_filtered_data_size(void);
 #endif
